+++ RNReactNativeHapticFeedback.podspec
@@ -1,6 +1,8 @@
 require 'json'
 package = JSON.parse(File.read(File.join(File.dirname(__FILE__), 'package.json')))
 
+folly_compiler_flags = '-DFOLLY_NO_CONFIG -DFOLLY_MOBILE=1 -DFOLLY_USE_LIBCPP=1 -Wno-comma -Wno-shorten-64-to-32'
+
 Pod::Spec.new do |s|
   s.name         = "RNReactNativeHapticFeedback"
   s.version      = package["version"]
@@ -14,10 +16,25 @@ Pod::Spec.new do |s|
   s.author             = { "author" => "m.kuczera@gmail.com" }
   s.platform     = :ios, "9.0"
   s.source       = { :git => "https://github.com/mkuczera/react-native-haptic-feedback.git", :tag => "master" }
-  s.source_files  = "ios/*.{h,m}"
+  s.source_files  = "ios/**/*.{h,m,mm,swift}"
   s.requires_arc = true
 
   s.dependency 'React-Core'
+
+  # This guard prevent to install the dependencies when we run `pod install` in the old architecture.
+  if ENV['RCT_NEW_ARCH_ENABLED'] == '1' then
+    s.compiler_flags = folly_compiler_flags + " -DRCT_NEW_ARCH_ENABLED=1"
+    s.pod_target_xcconfig    = {
+        "HEADER_SEARCH_PATHS" => "\"$(PODS_ROOT)/boost\"",
+        "CLANG_CXX_LANGUAGE_STANDARD" => "c++17"
+    }
+
+    s.dependency "React-Codegen"
+    s.dependency "RCT-Folly"
+    s.dependency "RCTRequired"
+    s.dependency "RCTTypeSafety"
+    s.dependency "ReactCommon/turbomodule/core"
+  end
 end
