+++ ios/RNHapticFeedback/DeviceUtils.h
