+++ ios/RNHapticFeedback/RNHapticFeedback.mm
@@ -1,5 +1,9 @@
+// Thanks to this guard, we won't import this header when we build for the old architecture.
+#ifdef RCT_NEW_ARCH_ENABLED
+#import "RNHapticFeedbackSpec.h"
+#endif
 
-#import "RNReactNativeHapticFeedback.h"
+#import "RNHapticFeedback.h"
 #import <UIKit/UIKit.h>
 #import <sys/utsname.h>
 #import "DeviceUtils.h"
@@ -7,10 +11,10 @@
 #import <AudioToolbox/AudioServices.h>
 
 static UISelectionFeedbackGenerator *selectionGenerator = nil;
-static NSMutableDictionary<NSNumber*, UIImpactFeedbackGenerator*> *impactGeneratorMap = nil;
+static NSMutableDictionary<NSString*, UIImpactFeedbackGenerator*> *impactGeneratorMap = nil;
 static UINotificationFeedbackGenerator *notificationGenerator = nil;
 
-@implementation RNReactNativeHapticFeedback
+@implementation RNHapticFeedback
 @synthesize bridge = _bridge;
 
 - (void)setBridge:(RCTBridge *)bridge
@@ -23,7 +27,7 @@ static UINotificationFeedbackGenerator *notificationGenerator = nil;
   return dispatch_get_main_queue();
 }
 
-RCT_EXPORT_MODULE();
+RCT_EXPORT_MODULE(HapticFeedback);
 
 RCT_EXPORT_METHOD(trigger:(NSString *)type options:(NSDictionary *)options)
 {
@@ -89,7 +93,7 @@ RCT_EXPORT_METHOD(trigger:(NSString *)type options:(NSDictionary *)options)
 }
 
 -(void)generateImpactFeedback:(UIImpactFeedbackStyle)style{
-    NSNumber *key = [NSNumber numberWithInteger: style];
+    NSString *key = [[NSNumber numberWithInteger: style] stringValue];
     if (impactGeneratorMap == nil)
         impactGeneratorMap = [[NSMutableDictionary alloc] init];
     if ([impactGeneratorMap objectForKey:key] == nil){
@@ -110,5 +114,13 @@ RCT_EXPORT_METHOD(trigger:(NSString *)type options:(NSDictionary *)options)
     [notificationGenerator prepare];
 }
 
+// Thanks to this guard, we won't compile this code when we build for the old architecture.
+#ifdef RCT_NEW_ARCH_ENABLED
+- (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
+    (const facebook::react::ObjCTurboModule::InitParams &)params
+{
+    return std::make_shared<facebook::react::NativeHapticFeedbackSpecJSI>(params);
+}
+#endif
 
 @end
