--- ios/RNReactNativeHapticFeedback.xcworkspace/contents.xcworkspacedata
@@ -1,9 +1,0 @@
-// !$*UTF8*$!
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "group:RNReactNativeHapticFeedback.xcodeproj">
-   </FileRef>
-</Workspace>
-
